requirements.txt
hachoir==3.1.2
Pillow==8.1.2
